Added delete confirm dialog

+++ src/app/settings/sports/sports.component.ts
@@ -6,6 +6,8 @@ import { Router } from '@angular/router';
 import { SportsService } from '../services/sports.service';
 import { Subscription } from 'rxjs';
 import { AddUpdateSportComponent } from './add-update-sport/add-update-sport.component';
+import { MatDialogRef } from '@angular/material';
+import { ConfirmDeleteDialogComponent } from 'src/app/shared/components/confirm-delete-dialog/confirm-delete-dialog.component';
 
 @Component({
   selector: 'app-sports',
@@ -21,6 +23,7 @@ export class SportsComponent implements OnInit, OnDestroy {
   selectedItems: string[];
 
   showPage = false;
+  dialogRef: MatDialogRef<ConfirmDeleteDialogComponent>;
 
   private sportsSubsription: Subscription;
 
@@ -36,9 +39,10 @@ export class SportsComponent implements OnInit, OnDestroy {
     this.sportsService.getSports();
     this.sportsSubsription = this.sportsService.getSportsUpdateListener()
       .subscribe((sports: Sport[]) => {
-        this.sportsList = new MatTableDataSource(sports)
+        this.sportsList = new MatTableDataSource(sports);
         this.sportsList.sort = this.sort;
         this.selectedItems = [];
+        this.showPage = true;
       });
   }
 
@@ -51,15 +55,15 @@ export class SportsComponent implements OnInit, OnDestroy {
   /** Selects all rows if they are not all selected; otherwise clear selection. */
   masterToggle() {
 
-    if(this.isAllSelected()) {
+    this.selectedItems = [];
+    if (this.isAllSelected()) {
       this.selection.clear();
-      this.selectedItems = [];
     } else {
       this.sportsList.data.forEach(row => {
-        this.selection.select(row)
+        this.selection.select(row);
         this.selectedItems.push(row.id);
       });
-    }        
+    }
   }
 
   rowClicked(item: Sport, wasChecked: boolean) {
@@ -71,11 +75,18 @@ export class SportsComponent implements OnInit, OnDestroy {
   }
 
   removeSelectedSports() {
+    this.dialogRef = this.dialog.open(ConfirmDeleteDialogComponent, {
+      disableClose: false
+    });
+    this.dialogRef.componentInstance.confirmMessage =
+      `Are you sure you want to delete ${this.selectedItems.length} item(s)?`;
 
-    // this.selectedItems.forEach(selectedSport => {
-    //   this.sportsList.data = this.sportsList.data.filter(sport => selectedSport.id !== sport.id);
-    // });
-    this.sportsService.deleteSports(this.selectedItems);
+    this.dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.sportsService.deleteSports(this.selectedItems);
+      }
+      this.dialogRef = null;
+    });
   }
 
   cancel() {
